d AC

+++ AtCoder/abc_105/main_d.cpp
@@ -7,14 +7,14 @@ typedef int64_t int64;
 typedef uint32_t uint;
 typedef uint64_t uint64;
 
-int pow2(uint e){
-    int p=1;
-    for(uint i=0; i<e; ++i){ p*=2; }
+int64 pow2(uint64 e){
+    int64 p=1;
+    for(uint64 i=0; i<e; ++i){ p*=2; }
     return p;
 }
-int nCr(int n, int r){
-    int ret=1;
-    for(int i=0; i<r; ++i){
+int64 nCr(int64 n, int64 r){
+    int64 ret=1;
+    for(int64 i=0; i<r; ++i){
         ret = ret * (n-i) / (i+1);
     }
     return ret;
@@ -24,18 +24,18 @@ int main(){
     ios_base::sync_with_stdio(false);
     cin.tie(NULL);
     
-    uint N, M; cin >> N >> M;
-    vector<uint> vA(N);
-    for(uint i=0; i<N; ++i){ cin >> vA[i]; }
+    uint64 N, M; cin >> N >> M;
+    vector<uint64> vA(N);
+    for(uint64 i=0; i<N; ++i){ cin >> vA[i]; }
     
     vector<uint64> vA_sum(N); vA_sum[0] = vA[0];
-    for(uint i=1; i<N; ++i){
+    for(uint64 i=1; i<N; ++i){
         vA_sum[i] = vA[i] + vA_sum[i-1];
     }
     
-    unordered_map<int, int> hashT;
-    for(uint i=0; i<N; ++i){
-        int mod = vA_sum[i] % M;
+    unordered_map<int64, int64> hashT;
+    for(uint64 i=0; i<N; ++i){
+        int64 mod = vA_sum[i] % M;
         
         auto itr = hashT.find( mod );
         if(itr!=hashT.end()){ ++(itr->second);
@@ -43,7 +43,7 @@ int main(){
         }
     }
     
-    int ans=0;
+    int64 ans=0;
     for(auto itr=hashT.begin(); itr!=hashT.end(); ++itr){
         if(itr->first!=0 && itr->second<2){ continue; }
